requirements.txt
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
pymongo==4.7.2
flask==3.0.3
flask-cors==4.0.0
matplotlib==3.9.0
seaborn==0.13.2
joblib==1.4.2
tqdm==4.66.4
python-dotenv==1.0.1
